Kenny-RAG/Kenny-RAG_v2.1/requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-huggingface
faiss-cpu
sentence-transformers
PyYAML
